:recycle: documents response
