+++ src/client/document/documents.rs
@@ -1,7 +1,7 @@
 use crate::query::query_string;
 use anyhow::anyhow;
-use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
+use crate::response::document::documents::GetDocumentsResponse;
 
 const RESOURCE: &str = "documents";
 const HEADER_KEY: &str = "X-Collaboflow-Authorization";
@@ -9,34 +9,6 @@ const HEADER_KEY: &str = "X-Collaboflow-Authorization";
 pub struct Documents {
     url: String,
     authorization_header: String,
-}
-
-#[derive(Debug, Deserialize, Clone, Serialize)]
-pub struct GetDocumentsResponse {
-    pub app_cd: i32,
-    pub processes_id: i32,
-    pub document_id: i32,
-    pub document_number: String,
-    pub title: String,
-    pub request_username: String,
-    pub request_usercd: String,
-    pub request_groupname: String,
-    pub request_groupcd: String,
-    pub request_titles: Vec<RequestTitles>,
-    pub represent_username: String,
-    pub represent_usercd: String,
-    pub flow_status: String,
-    pub request_date: String,
-    pub end_date: String,
-    pub link: String,
-    pub error: bool,
-}
-
-#[derive(Debug, Deserialize, Clone, Serialize)]
-pub struct RequestTitles {
-    id: String,
-    code: String,
-    name: String,
 }
 
 impl Documents {
